+++ assets/empty_packs/MaterialRounded.dart
@@ -1,3 +1,3 @@
-import 'package:flutter/material.dart';
+import 'package:flutter_iconpicker/Models/icon_picker_icon.dart';
 
-Map<String, IconData> roundedIcons = {};
+Map<String, IconPickerIcon> roundedIcons = {};