+++ lib/IconPicker/icon_picker.dart
@@ -3,17 +3,23 @@
 /// https://github.com/Ahmadre
 /// rebar.ahmad@gmail.com
 
+import 'dart:async';
+
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_iconpicker/controllers/icon_controller.dart';
 import 'package:provider/provider.dart';
+import 'package:scrollview_observer/scrollview_observer.dart';
+import '../Models/icon_picker_icon.dart';
 import 'icons.dart';
 import '../Models/icon_pack.dart';
 import '../Helpers/color_brightness.dart';
 
 class FIPIconPicker extends StatefulWidget {
   final FIPIconController iconController;
+  final Color? selectedIconBackgroundColor;
   final List<IconPack>? iconPack;
-  final Map<String, IconData>? customIconPack;
+  final Map<String, IconPickerIcon>? customIconPack;
   final double? iconSize;
   final Color? iconColor;
   final String? noResultsText;
@@ -29,6 +35,7 @@ class FIPIconPicker extends StatefulWidget {
     required this.iconSize,
     required this.noResultsText,
     required this.backgroundColor,
+    this.selectedIconBackgroundColor,
     this.mainAxisSpacing,
     this.crossAxisSpacing,
     this.iconColor,
@@ -41,19 +48,52 @@ class FIPIconPicker extends StatefulWidget {
 }
 
 class _FIPIconPickerState extends State<FIPIconPicker> {
+  ScrollController scrollController = ScrollController();
+  late GridObserverController observerController =
+      GridObserverController(controller: scrollController);
+
+  bool scrolledToSelectedIcon = false;
+  int selectedIconIndex = -1;
+
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
       if (widget.customIconPack != null) {
         if (mounted) widget.iconController.addAll(widget.customIconPack ?? {});
       }
 
-      if (widget.iconPack != null)
+      if (widget.iconPack != null) {
         for (var pack in widget.iconPack!) {
-          if (mounted)
+          if (mounted) {
             widget.iconController.addAll(FIPIconManager.getSelectedPack(pack));
+          }
         }
+        if (mounted &&
+            !scrolledToSelectedIcon &&
+            widget.iconController.shouldScrollToSelectedIcon &&
+            widget.iconController.isSelectedIconAvailable &&
+            widget.iconController.entries.firstWhereOrNull((item) =>
+                    widget.iconController.selectedIcon == item.value) !=
+                null) {
+          Timer.periodic(const Duration(milliseconds: 100), (timer) async {
+            if (scrollController.hasClients) {
+              scrolledToSelectedIcon = true;
+              selectedIconIndex = widget.iconController.entries
+                  .toList()
+                  .indexWhere((item) =>
+                      widget.iconController.selectedIcon == item.value);
+              setState(() {});
+              await observerController.jumpTo(
+                index: selectedIconIndex,
+                offset: (_) => 24,
+              );
+              timer.cancel();
+            }
+          });
+        }
+      }
     });
   }
 
@@ -96,36 +136,58 @@ class _FIPIconPickerState extends State<FIPIconPicker> {
               _getListEmptyMsg()
             else
               Positioned.fill(
-                child: GridView.builder(
-                    itemCount: controller.length,
-                    gridDelegate: SliverGridDelegateWithMaxCrossAxisExtent(
-                      childAspectRatio: 1 / 1,
-                      mainAxisSpacing: widget.mainAxisSpacing ?? 5,
-                      crossAxisSpacing: widget.crossAxisSpacing ?? 5,
-                      maxCrossAxisExtent:
-                          widget.iconSize != null ? widget.iconSize! + 10 : 50,
-                    ),
-                    itemBuilder: (context, index) {
-                      var item = controller.entries.elementAt(index);
+                child: GridViewObserver(
+                  controller: observerController,
+                  child: GridView.builder(
+                      controller: scrollController,
+                      itemCount: controller.length,
+                      gridDelegate: SliverGridDelegateWithMaxCrossAxisExtent(
+                        childAspectRatio: 1 / 1,
+                        mainAxisSpacing: widget.mainAxisSpacing ?? 5,
+                        crossAxisSpacing: widget.crossAxisSpacing ?? 5,
+                        maxCrossAxisExtent: widget.iconSize != null
+                            ? widget.iconSize! + 10
+                            : 50,
+                      ),
+                      itemBuilder: (context, index) {
+                        final MapEntry<String, IconPickerIcon> item =
+                            controller.entries.elementAt(index);
 
-                      return GestureDetector(
-                        onTap: () => Navigator.pop(context, item.value),
-                        child: widget.showTooltips!
-                            ? Tooltip(
-                                message: item.key,
-                                child: Icon(
-                                  item.value,
-                                  size: widget.iconSize,
-                                  color: widget.iconColor,
-                                ),
-                              )
-                            : Icon(
-                                item.value,
-                                size: widget.iconSize,
-                                color: widget.iconColor,
-                              ),
-                      );
-                    }),
+                        final isSelectedIcon = widget
+                                .iconController.isSelectedIconAvailable &&
+                            widget.iconController.selectedIcon! == item.value;
+
+                        final selectedIconColor =
+                            widget.selectedIconBackgroundColor ??
+                                (Theme.of(context).brightness == Brightness.dark
+                                    ? Colors.grey[800]
+                                    : Colors.grey[400]);
+
+                        return ClipRRect(
+                          borderRadius: BorderRadius.circular(8),
+                          child: Container(
+                            color: isSelectedIcon ? selectedIconColor : null,
+                            child: GestureDetector(
+                              onTap: () => Navigator.pop(context, item.value),
+                              child: widget.showTooltips!
+                                  ? Tooltip(
+                                      message: item.key,
+                                      child: Icon(
+                                        item.value.data,
+                                        size: widget.iconSize,
+                                        color: widget.iconColor,
+                                      ),
+                                    )
+                                  : Icon(
+                                      item.value.data,
+                                      size: widget.iconSize,
+                                      color: widget.iconColor,
+                                    ),
+                            ),
+                          ),
+                        );
+                      }),
+                ),
               ),
             IgnorePointer(
               child: Container(