+++ lib/IconPicker/icons.dart
@@ -15,15 +15,14 @@ import 'package:flutter_iconpicker/IconPicker/Packs/MaterialSharp.dart';
 
 import '../Models/icon_pack.dart';
 
+import '../Models/icon_picker_icon.dart';
 import 'Packs/Material.dart';
 import 'Packs/Cupertino.dart';
 import 'Packs/FontAwesome.dart';
 import 'Packs/LineIcons.dart';
 
-import 'package:flutter/widgets.dart';
-
 class FIPIconManager {
-  static Map<String, IconData> getSelectedPack(IconPack? pickedPack) {
+  static Map<String, IconPickerIcon> getSelectedPack(IconPack? pickedPack) {
     switch (pickedPack) {
       case IconPack.material:
         return defaultIcons;
@@ -42,7 +41,7 @@ class FIPIconManager {
       case IconPack.lineAwesomeIcons:
         return lineAwesomeIcons;
       default:
-        return <String, IconData>{};
+        return <String, IconPickerIcon>{};
     }
   }
 }