+++ lib/controllers/icon_controller.dart
@@ -1,13 +1,34 @@
 import 'package:flutter/material.dart';
 
+import '../Models/icon_picker_icon.dart';
+
 class FIPIconController with ChangeNotifier {
-  FIPIconController();
+  FIPIconController({
+    required bool shouldScrollToSelectedIcon,
+    IconPickerIcon? selectedIcon,
+  })  : _selectedIcon = selectedIcon,
+        _shouldScrollToSelectedIcon = shouldScrollToSelectedIcon;
 
-  Map<String, IconData> _icons = {};
+  bool _shouldScrollToSelectedIcon;
 
-  Map<String, IconData> get icons => _icons;
+  bool get shouldScrollToSelectedIcon => _shouldScrollToSelectedIcon;
 
-  set icons(Map<String, IconData> val) {
+  IconPickerIcon? _selectedIcon;
+
+  IconPickerIcon? get selectedIcon => _selectedIcon;
+
+  set selectedIcon(IconPickerIcon? val) {
+    _selectedIcon = val;
+    notifyListeners();
+  }
+
+  bool get isSelectedIconAvailable => _selectedIcon != null;
+
+  Map<String, IconPickerIcon> _icons = {};
+
+  Map<String, IconPickerIcon> get icons => _icons;
+
+  set icons(Map<String, IconPickerIcon> val) {
     _icons = val;
     notifyListeners();
   }
@@ -21,11 +42,11 @@ class FIPIconController with ChangeNotifier {
     notifyListeners();
   }
 
-  get length => _icons.length;
+  int get length => _icons.length;
 
-  get entries => _icons.entries;
+  Iterable<MapEntry<String, IconPickerIcon>> get entries => _icons.entries;
 
-  void addAll(Map<String, IconData> pack) {
+  void addAll(Map<String, IconPickerIcon> pack) {
     _icons.addAll(pack);
     notifyListeners();
   }